% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencyconverter.R
\name{convert_currency}
\alias{convert_currency}
\title{Function that queries historical exchange rates for any currency}
\usage{
convert_currency(base = "EUR", symbol = "USD", days = 2)
}
\arguments{
\item{base}{string}

\item{symbol}{string}

\item{days}{number}
}
\value{
data.frame of historic exchange rates
}
\description{
Function that queries historical exchange rates for any currency
}
\examples{
convert_currency(base = 'EUR', symbol = 'USD', days = 2)
}
