% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{forint}
\alias{forint}
\title{Formatter function for Hungarian Forint}
\usage{
forint(x)
}
\arguments{
\item{x}{numer}
}
\value{
string
}
\description{
Formatter function for Hungarian Forint
}
\examples{
forint(100000)
forint(10.3232432)
}
